Move Language Service contracts over to Language Service project
