+++ src/Microsoft.SqlTools.LanguageService/LanguageServices/Completion/Extension/CompletionExtensionParams.cs
@@ -1,4 +1,4 @@
-﻿//
+//
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
@@ -8,7 +8,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Microsoft.SqlTools.ServiceLayer.LanguageServices.Completion.Extension
+namespace Microsoft.SqlTools.LanguageService.LanguageServices.Completion.Extension
 {
 
     [Serializable]
