+++ src/Microsoft.SqlTools.LanguageService/LanguageServices/Contracts/IntelliSenseReady.cs
@@ -7,7 +7,7 @@
 
 using Microsoft.SqlTools.Hosting.Protocol.Contracts;
 
-namespace Microsoft.SqlTools.ServiceLayer.LanguageServices.Contracts
+namespace Microsoft.SqlTools.LanguageService.LanguageServices.Contracts
 {
     /// <summary>
     /// Parameters sent back with an IntelliSense ready event
