+++ src/Microsoft.SqlTools.LanguageService/LanguageServices/Contracts/NonTSqlNotification.cs
@@ -8,7 +8,7 @@
 using System.Collections.Generic;
 using Microsoft.SqlTools.Hosting.Protocol.Contracts;
 
-namespace Microsoft.SqlTools.ServiceLayer.LanguageServices.Contracts
+namespace Microsoft.SqlTools.LanguageService.LanguageServices.Contracts
 {
     /// <summary>
     /// Parameters to be sent back with a non-TSQL file event.
