+++ src/Microsoft.SqlTools.LanguageService/LanguageServices/Contracts/SyntaxParse.cs
@@ -7,7 +7,7 @@
 
 using Microsoft.SqlTools.Hosting.Protocol.Contracts;
 
-namespace Microsoft.SqlTools.ServiceLayer.LanguageServices.Contracts
+namespace Microsoft.SqlTools.LanguageService.LanguageServices.Contracts
 {
     public class SyntaxParseParams 
     {
